add base class for models that implements iNotifyPropertyChanged
